requirements.txt
langchain
langchain_community
langchain_groq
ipykernel
python-dotenv
requests
replicate
pillow
rembg
onnxruntime